fix: Ensures that sidebar and topbar do not flicker when navigating pages
