+++ src/app/(main)/layout.tsx
@@ -1,0 +1,14 @@
+import Sidebar from "@/components/navigation/Sidebar";
+import TopNavBar from "@/components/navigation/TopNavigationBar";
+
+export default function MainLayout({ children }: { children: React.ReactNode }) {
+  return (
+    <div className="flex h-screen bg-background flex-col overflow-hidden">
+      <TopNavBar />
+      <div className="flex flex-row flex-1 w-full overflow-hidden">
+        <Sidebar />
+        {children}
+      </div>
+    </div>
+  );
+}
